+++ org.eclipse.dawnsci.analysis.dataset/src/org/eclipse/dawnsci/analysis/dataset/impl/CompoundFloatDataset.java
@@ -1254,7 +1254,7 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 		if (bds.getSize() == 1) {
 			final IndexIterator it = getIterator();
 			if (is == 1) {
-				final float db = (float) bds.getElementDoubleAbs(0); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
+				final double db = bds.getElementDoubleAbs(0);
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						data[it.index + i] += db;
@@ -1263,7 +1263,7 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						data[it.index + i] += (float) bds.getElementDoubleAbs(i); // GET_ELEMENT_WITH_CAST
+						data[it.index + i] += bds.getElementDoubleAbs(i);
 					}
 				}
 			} else {
@@ -1275,16 +1275,16 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 			if (is == 1) {
 				while (it.hasNext()) {
 					final double db = it.bDouble;
-					data[it.aIndex] += (float) db; // ADD_CAST
+					data[it.aIndex] += db;
 					for (int i = 1; i < isize; i++) {
-						data[it.aIndex + i] += (float) db; // ADD_CAST
+						data[it.aIndex + i] += db;
 					}
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
-					data[it.aIndex] += (float) it.bDouble; // ADD_CAST
+					data[it.aIndex] += it.bDouble;
 					for (int i = 1; i < isize; i++) {
-						data[it.aIndex + i] += (float) bds.getElementDoubleAbs(it.bIndex + i); // GET_ELEMENT_WITH_CAST
+						data[it.aIndex + i] += bds.getElementDoubleAbs(it.bIndex + i);
 					}
 				}
 			} else {
@@ -1302,19 +1302,16 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 		if (bds.getSize() == 1) {
 			final IndexIterator it = getIterator();
 			if (is == 1) {
-				final float db = (float) bds.getElementDoubleAbs(0); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
-				// final double db = bds.getElementDoubleAbs(0); // INT_USE
+				final double db = bds.getElementDoubleAbs(0);
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						data[it.index + i] -= db;
-						// data[it.index + i] = (float) (data[it.index + i] - db); // INT_USE // ADD_CAST
 					}
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						data[it.index + i] -= (float) bds.getElementDoubleAbs(i); // GET_ELEMENT_WITH_CAST
-						// data[it.index + i] = (float) (data[it.index + i] - bds.getElementDoubleAbs(i)); // INT_USE // ADD_CAST
+						data[it.index + i] -= bds.getElementDoubleAbs(i);
 					}
 				}
 			} else {
@@ -1327,17 +1324,14 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 				while (it.hasNext()) {
 					final double db = it.bDouble;
 					for (int i = 0; i < isize; i++) {
-						data[it.aIndex + i] -= (float) db; // ADD_CAST
-						// data[it.aIndex + i] = (float) (it.aDouble - db); // INT_USE // ADD_CAST
+						data[it.aIndex + i] -= db;
 					}
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
-					data[it.aIndex] -= (float) it.bDouble; // ADD_CAST
-					// data[it.aIndex] = (float) (it.aDouble - it.bDouble); // INT_USE // ADD_CAST
+					data[it.aIndex] -= it.bDouble;
 					for (int i = 1; i < isize; i++) {
-						data[it.aIndex + i] -= (float) bds.getElementDoubleAbs(it.bIndex + i); // GET_ELEMENT_WITH_CAST
-						// data[it.aIndex + i] = (float) (data[it.aIndex + i] - bds.getElementDoubleAbs(it.bIndex + i)); // INT_USE // ADD_CAST
+						data[it.aIndex + i] -= bds.getElementDoubleAbs(it.bIndex + i);
 					}
 				}
 			} else {
@@ -1355,7 +1349,7 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 		if (bds.getSize() == 1) {
 			final IndexIterator it = getIterator();
 			if (is == 1) {
-				final float db = (float) bds.getElementDoubleAbs(0); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
+				final double db = bds.getElementDoubleAbs(0);
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						data[it.index + i] *= db;
@@ -1364,7 +1358,7 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						data[it.index + i] *= (float) bds.getElementDoubleAbs(i); // GET_ELEMENT_WITH_CAST
+						data[it.index + i] *= bds.getElementDoubleAbs(i);
 					}
 				}
 			} else {
@@ -1377,14 +1371,14 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 				while (it.hasNext()) {
 					final double db = it.bDouble;
 					for (int i = 0; i < isize; i++) {
-						data[it.aIndex + i] *= (float) db; // ADD_CAST
+						data[it.aIndex + i] *= db;
 					}
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
-					data[it.aIndex] *= (float) it.bDouble; // ADD_CAST
+					data[it.aIndex] *= it.bDouble;
 					for (int i = 1; i < isize; i++) {
-						data[it.aIndex + i] *= (float) bds.getElementDoubleAbs(it.bIndex + i); // GET_ELEMENT_WITH_CAST
+						data[it.aIndex + i] *= bds.getElementDoubleAbs(it.bIndex + i);
 					}
 				}
 			} else {
@@ -1402,29 +1396,21 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 		if (bds.getSize() == 1) {
 			final IndexIterator it = getIterator();
 			if (is == 1) {
-				final float db = (float) bds.getElementDoubleAbs(0); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
-				// final double db = bds.getElementDoubleAbs(0); // INT_USE
+				final double db = bds.getElementDoubleAbs(0);
 				// if (db == 0) { // INT_USE
 				// 	fill(0); // INT_USE
 				// } else { // INT_USE
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						data[it.index + i] /= db;
-						// data[it.index + i] = (float) (data[it.index + i] / db); // INT_USE // ADD_CAST
 					}
 				}
 				// } // INT_USE
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						final float db = (float) bds.getElementDoubleAbs(i); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
-						// final double db = bds.getElementDoubleAbs(0); // INT_USE
-						data[it.index + i] /= db;
-						// try { // INT_USE
-						// 	data[it.index + i] = (float) (data[it.index + i] / db); // INT_USE // ADD_CAST
-						// } catch (ArithmeticException e) { // INT_USE
-						// 	data[it.index + i] = 0; // INT_USE
-						// } // INT_USE
+						final double db = bds.getElementDoubleAbs(i);
+						data[it.index + i] /= db; // INT_EXCEPTION
 					}
 				}
 			} else {
@@ -1442,22 +1428,15 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 					// 	}// INT_USE
 					// } else { // INT_USE
 					for (int i = 0; i < isize; i++) {
-						data[it.aIndex + i] /= (float) db; // ADD_CAST
-					// 	data[it.aIndex + i] = (float) (data[it.aIndex + i] / db); // INT_USE // ADD_CAST
+						data[it.aIndex + i] /= db;
 					}
 					// } // INT_USE
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						final float db = (float) bds.getElementDoubleAbs(it.bIndex + i); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
-						// final double db = bds.getElementDoubleAbs(0); // INT_USE
-						data[it.aIndex + i] /= db;
-						// try { // INT_USE
-						// 	data[it.aIndex + i] = (float) (data[it.aIndex + i] / db); // INT_USE // ADD_CAST
-						// } catch (ArithmeticException e) { // INT_USE
-						// 	data[it.aIndex + i] = 0; // INT_USE
-						// } // INT_USE
+						final double db = bds.getElementDoubleAbs(it.bIndex + i);
+						data[it.aIndex + i] /= db; // INT_EXCEPTION
 					}
 				}
 			} else {
@@ -1486,23 +1465,20 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 		if (bds.getSize() == 1) {
 			final IndexIterator it = getIterator();
 			if (is == 1) {
-				final float db = (float) bds.getElementDoubleAbs(0); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
-				// final double db = bds.getElementDoubleAbs(0); // INT_USE
+				final double db = bds.getElementDoubleAbs(0);
 				// if (db == 0) { // INT_USE
 				// 	fill(0); // INT_USE
 				// } else { // INT_USE
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						data[it.index + i] %= db;
-						// data[it.index + i] = (float) (data[it.index + i] % db); // INT_USE // ADD_CAST
 					}
 				}
 				// } // INT_USE
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						data[it.index + i] %= (float) bds.getElementDoubleAbs(i); // GET_ELEMENT_WITH_CAST
-						// data[it.index + i] = (float) (data[it.index + i] % bds.getElementDoubleAbs(i)); // INT_USE // ADD_CAST
+						data[it.index + i] %= bds.getElementDoubleAbs(i); // INT_EXCEPTION
 					}
 				}
 			} else {
@@ -1520,20 +1496,13 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 					// } else { // INT_USE
 					for (int i = 0; i < isize; i++)
 						data[it.aIndex + i] %= db;
-					// 	data[it.aIndex + i] = (float) (data[it.aIndex + i] % db); // INT_USE // ADD_CAST
 					// } // INT_USE
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
-						final float db = (float) bds.getElementDoubleAbs(it.bIndex + i); // PRIM_TYPE // GET_ELEMENT_WITH_CAST
-						// final double db = bds.getElementDoubleAbs(0); // INT_USE
-						data[it.aIndex + i] %= db;
-						// try { // INT_USE
-						// 	data[it.aIndex + i] = (float) (data[it.aIndex + i] % db); // INT_USE // ADD_CAST
-						// } catch (ArithmeticException e) { // INT_USE
-						// 	data[it.aIndex + i] = 0; // INT_USE
-						// } // INT_USE
+						final double db = bds.getElementDoubleAbs(it.bIndex + i);
+						data[it.aIndex + i] %= db; // INT_EXCEPTION
 					}
 				}
 			} else {
@@ -1557,11 +1526,11 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 					while (it.hasNext()) {
 						for (int i = 0; i < isize; i++) {
 							final double v = Math.pow(data[it.index + i], vr);
-							// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_ZEROTEST
-							// 	data[it.index + i] = 0; // INT_ZEROTEST
-							// } else { // INT_ZEROTEST
+							// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_USE
+							// 	data[it.index + i] = 0; // INT_USE
+							// } else { // INT_USE
 							data[it.index + i] = (float) v; // PRIM_TYPE_LONG // ADD_CAST
-							// } // INT_ZEROTEST
+							// } // INT_USE
 						}
 					}
 				} else {
@@ -1570,11 +1539,11 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 						for (int i = 0; i < isize; i++) {
 							Complex zd = new Complex(data[it.index + i], 0.);
 							final double v = zd.pow(zv).getReal();
-							// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_ZEROTEST
-							// 	data[it.index + i] = 0; // INT_ZEROTEST
-							// } else { // INT_ZEROTEST
+							// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_USE
+							// 	data[it.index + i] = 0; // INT_USE
+							// } else { // INT_USE
 							data[it.index + i] = (float) v; // PRIM_TYPE_LONG // ADD_CAST
-							// } // INT_ZEROTEST
+							// } // INT_USE
 						}
 					}
 				}
@@ -1582,22 +1551,22 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						final double v = Math.pow(data[it.index + i], vr);
-						// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_ZEROTEST
-						// 	data[it.index + i] = 0; // INT_ZEROTEST
-						// } else { // INT_ZEROTEST
+						// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_USE
+						// 	data[it.index + i] = 0; // INT_USE
+						// } else { // INT_USE
 						data[it.index + i] = (float) v; // PRIM_TYPE_LONG // ADD_CAST
-						// } // INT_ZEROTEST
+						// } // INT_USE
 					}
 				}
 			} else if (is == isize) {
 				while (it.hasNext()) {
 					for (int i = 0; i < isize; i++) {
 						final double v = Math.pow(data[it.index + i], bds.getElementDoubleAbs(i));
-						// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_ZEROTEST
-						// 	data[it.index + i] = 0; // INT_ZEROTEST
-						// } else { // INT_ZEROTEST
+						// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_USE
+						// 	data[it.index + i] = 0; // INT_USE
+						// } else { // INT_USE
 						data[it.index + i] = (float) v; // PRIM_TYPE_LONG // ADD_CAST
-						// } // INT_ZEROTEST
+						// } // INT_USE
 					}
 				}
 			}
@@ -1606,18 +1575,18 @@ public class CompoundFloatDataset extends AbstractCompoundDataset {
 			it.setOutputDouble(true);
 			while (it.hasNext()) {
 				double v = Math.pow(it.aDouble, it.bDouble);
-				// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_ZEROTEST
-				// 	data[it.aIndex] = 0; // INT_ZEROTEST
-				// } else { // INT_ZEROTEST
+				// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_USE
+				// 	data[it.aIndex] = 0; // INT_USE
+				// } else { // INT_USE
 				data[it.aIndex] = (float) v; // PRIM_TYPE_LONG // ADD_CAST
-				// } // INT_ZEROTEST
+				// } // INT_USE
 				for (int i = 1; i < isize; i++) {
 					v = Math.pow(data[it.aIndex + i], bds.getElementDoubleAbs(it.bIndex + i));
-					// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_ZEROTEST
-					// 	data[it.aIndex + i] = 0; // INT_ZEROTEST
-					// } else { // INT_ZEROTEST
+					// if (Double.isInfinite(v) || Double.isNaN(v)) { // INT_USE
+					// 	data[it.aIndex + i] = 0; // INT_USE
+					// } else { // INT_USE
 					data[it.aIndex + i] = (float) v; // PRIM_TYPE_LONG // ADD_CAST
-					// } // INT_ZEROTEST
+					// } // INT_USE
 				}
 			}
 		}
